[10.0][IMP] web_switch_company_warning module
[FIX] for enterprise Odoo versions

[IMP] add user switch detection and change module name
