+++ web_switch_context_warning/README.rst
@@ -2,12 +2,13 @@
    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
    :alt: License: AGPL-3
 
-===========================
-Web Switch Company Warning
-===========================
+==========================
+Web Switch Context Warning
+==========================
 
 
-Shows a warning if current company has been switched in another tab or window.
+Shows a warning if current user, company or database have been switched
+in another tab or window.
 
 
 Known issues / Roadmap
@@ -16,7 +17,7 @@ Known issues / Roadmap
 
     * If the browser don't implement Sharded Worker (http://www.w3.org/TR/workers/#sharedworker), the warning message will not be displayed (there is no polyfill).
 
-    * Switching company in a separate browser or in private browsing mode will not be detected by this module. It's a limitation of Shared Worker(limit to browser session, server:port...)
+    * Switching user, database or company in a separate browser or in private browsing mode will not be detected by this module. It's a limitation of Shared Worker(limit to browser session, server:port...)
 
 
 Bug Tracker
