+++ web_switch_context_warning/__manifest__.py
@@ -2,12 +2,12 @@
 # Copyright 2017 Akretion <raphael.reverdy@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
-    "name": "Multicompany - Switch Company Warning",
-    "summary": "Show a warning if current company has been switched"
-               " in another tab or window.",
-    "version": "10.0.0.1.1",
+    "name": "Switch Context Warning",
+    "summary": "Show a warning if current user, company or database"
+               " have been switched in another tab or window.",
+    "version": "10.0.1.2.0",
     "category": "web",
-    "website": "http://akretion.com",
+    "website": "https://github.com/OCA/web",
     "license": "AGPL-3",
     "author": "Akretion, "
               "Odoo Community Association (OCA)",
@@ -18,7 +18,7 @@
         "view/view.xml",
     ],
     "qweb": [
-        "static/src/xml/switch_company_warning.xml",
+        "static/src/xml/switch_context_warning.xml",
     ],
     'installable': True,
     "application": False,
