+++ web_switch_context_warning/static/src/xml/switch_company_warning.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <template>
-    <t t-name="web_switch_company_warning.warningWidget">
+    <t t-name="web_switch_context_warning.warningWidget">
         <div class="container-fluid bg-warning" style="text-align: center; display:none;">
-            <h3>You switched to a different company or database with another tab or window</h3>
+            <h3>You switched to a different user, company or database with another tab or window</h3>
             <p><button onclick="location.reload(true);" class="btn">Reload</button> to refresh your session</p>
         </div>
     </t>
